mux: Move filtering from tree to mux
The filtering is not useful for basic params, let's move it to mux and
while on it also remove the avocado arguments (options) to the
yaml_to_mux plugin.
