+++ avocado/core/mux.py
@@ -99,3 +99,63 @@ class MuxTreeNodeDebug(MuxTreeNode, tree.TreeNodeDebug):
         MuxTreeNode.__init__(self, name, value, parent, children)
         tree.TreeNodeDebug.__init__(self, name, value, parent, children,
                                     srcyaml)
+
+
+#
+# Tree filtering
+#
+def path_parent(path):
+    """
+    From a given path, return its parent path.
+
+    :param path: the node path as string.
+    :return: the parent path as string.
+    """
+    parent = path.rpartition('/')[0]
+    if not parent:
+        return '/'
+    return parent
+
+
+def apply_filters(root, filter_only=None, filter_out=None):
+    """
+    Apply a set of filters to the tree.
+
+    The basic filtering is filter only, which includes nodes,
+    and the filter out rules, that exclude nodes.
+
+    Note that filter_out is stronger than filter_only, so if you filter out
+    something, you could not bypass some nodes by using a filter_only rule.
+
+    :param filter_only: the list of paths which will include nodes.
+    :param filter_out: the list of paths which will exclude nodes.
+    :return: the original tree minus the nodes filtered by the rules.
+    """
+    if filter_only is None:
+        filter_only = []
+    else:
+        filter_only = [_.rstrip('/') for _ in filter_only if _]
+    if filter_out is None:
+        filter_out = []
+    else:
+        filter_out = [_.rstrip('/') for _ in filter_out if _]
+    for node in root.iter_children_preorder():
+        keep_node = True
+        for path in filter_only:
+            if path == '':
+                continue
+            if node.path == path:
+                keep_node = True
+                break
+            if node.parent and node.parent.path == path_parent(path):
+                keep_node = False
+                continue
+        for path in filter_out:
+            if path == '':
+                continue
+            if node.path == path:
+                keep_node = False
+                break
+        if not keep_node:
+            node.detach()
+    return root
