+++ avocado/plugins/run.py
@@ -132,18 +132,6 @@ class Run(CLICmd):
 
         loader.add_loader_options(parser)
 
-        mux = parser.add_argument_group('test parameters')
-        mux.add_argument('--filter-only', nargs='*', default=[],
-                         help='Filter only path(s) from multiplexing')
-        mux.add_argument('--filter-out', nargs='*', default=[],
-                         help='Filter out path(s) from multiplexing')
-        mux.add_argument('--mux-path', nargs='*', default=None,
-                         help="List of paths used to determine path "
-                         "priority when querying for parameters")
-        mux.add_argument('--mux-inject', default=[], nargs='*',
-                         help="Inject [path:]key:node values into the "
-                         "final multiplex tree.")
-
     def run(self, args):
         """
         Run test modules or simple tests.
