+++ avocado/plugins/yaml_to_mux.py
@@ -249,10 +249,28 @@ class YamlToMux(CLI):
             mux.add_argument("-m", "--mux-yaml", nargs='*', metavar="FILE",
                              help="Location of one or more Avocado"
                              " multiplex (.yaml) FILE(s) (order dependent)")
+            mux.add_argument('--mux-filter-only', nargs='*', default=[],
+                             help='Filter only path(s) from multiplexing')
+            mux.add_argument('--mux-filter-out', nargs='*', default=[],
+                             help='Filter out path(s) from multiplexing')
+            mux.add_argument('--mux-path', nargs='*', default=None,
+                             help="List of paths used to determine path "
+                             "priority when querying for parameters")
+            mux.add_argument('--mux-inject', default=[], nargs='*',
+                             help="Inject [path:]key:node values into the "
+                             "final multiplex tree.")
+            mux = subparser.add_argument_group("yaml to mux options "
+                                               "[deprecated]")
             mux.add_argument("--multiplex", nargs='*',
                              default=None, metavar="FILE",
                              help="DEPRECATED: Location of one or more Avocado"
                              " multiplex (.yaml) FILE(s) (order dependent)")
+            mux.add_argument("--filter-only", nargs='*', default=[],
+                             help="DEPRECATED: Filter only path(s) from "
+                             "multiplexing (use --mux-only instead)")
+            mux.add_argument("--filter-out", nargs='*', default=[],
+                             help="DEPRECATED: Filter out path(s) from "
+                             "multiplexing (use --mux-out instead)")
 
     @staticmethod
     def _log_deprecation_msg(deprecated, current):
@@ -263,12 +281,34 @@ class YamlToMux(CLI):
         logging.getLogger("avocado.app").warning(msg, deprecated, current)
 
     def run(self, args):
+        # Deprecated filters
+        only = getattr(args, "filter_only", None)
+        if only:
+            self._log_deprecation_msg("--filter-only", "--mux-only")
+            mux_filter_only = getattr(args, "mux_filter_only")
+            if mux_filter_only:
+                args.mux_filter_only = mux_filter_only + only
+            else:
+                args.mux_filter_only = only
+        out = getattr(args, "filter_out", None)
+        if out:
+            self._log_deprecation_msg("--filter-out", "--mux-out")
+            mux_filter_out = getattr(args, "mux_filter_out")
+            if mux_filter_out:
+                args.mux_filter_out = mux_filter_out + out
+            else:
+                args.mux_filter_out = out
+        if args.xxx_variants.debug:
+            data = mux.MuxTreeNodeDebug()
+        else:
+            data = mux.MuxTreeNode()
+
         # Merge the multiplex
         multiplex_files = getattr(args, "mux_yaml", None)
         if multiplex_files:
             debug = getattr(args, "mux_debug", False)
             try:
-                args.xxx_variants.data_merge(create_from_yaml(multiplex_files, debug))
+                data.merge(create_from_yaml(multiplex_files, debug))
             except IOError as details:
                 logging.getLogger("avocado.app").error(details.strerror)
                 sys.exit(exit_codes.AVOCADO_JOB_FAIL)
@@ -279,7 +319,24 @@ class YamlToMux(CLI):
             self._log_deprecation_msg("--multiplex", "--mux-yaml")
             debug = getattr(args, "mux_debug", False)
             try:
+                data.merge(create_from_yaml(multiplex_files, debug))
                 args.xxx_variants.data_merge(create_from_yaml(multiplex_files, debug))
             except IOError as details:
                 logging.getLogger("avocado.app").error(details.strerror)
                 sys.exit(exit_codes.AVOCADO_JOB_FAIL)
+
+        # Extend default multiplex tree of --mux-inject values
+        for inject in getattr(args, "mux_inject", []):
+            entry = inject.split(':', 3)
+            if len(entry) < 2:
+                raise ValueError("key:entry pairs required, found only %s"
+                                 % (entry))
+            elif len(entry) == 2:   # key, entry
+                entry.insert(0, '')  # add path='' (root)
+            data.get_node(entry[0], True).value[entry[1]] = entry[2]
+
+        mux_filter_only = getattr(args, 'mux_filter_only', None)
+        mux_filter_out = getattr(args, 'mux_filter_out', None)
+        data = mux.apply_filters(data, mux_filter_only, mux_filter_out)
+        if data != mux.MuxTreeNode():
+            args.xxx_variants.data_merge(data)
